# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Followup API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts",
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "add new post",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/corporate-posts": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "corporate-posts",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/update-post/{post}": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "update-post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/delete-post/{post}": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "delete-post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/user/{user}/posts": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "user posts",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
    },
...
}